requirements.txt
# Azure Services
azure-functions>=1.17.0
azure-eventhub>=5.11.0
azure-identity>=1.15.0
azure-servicebus>=7.11.0
azure-monitor-opentelemetry>=1.2.0
azure-cosmos>=4.5.1

# AI/ML
langchain>=0.1.0
openai>=1.12.0

# Telemetry & Monitoring
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation>=0.44b0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Configuration & Environment
python-dotenv>=1.0.0

# Utilities
pandas>=2.2.0
numpy>=1.24.0
pydantic>=2.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Logging
structlog>=24.1.0

# Development
black>=24.1.0
flake8>=7.0.0
isort>=5.13.0 